Override some highlights in Vim
Simplifies Visual and Folded highlight groups. The syntax colors in
visual select was too distracting, and the Folded highlight group stood out
too much.

`colorscheme default` is necessary for the autocmd to run on startup,
since the colorscheme is not set anywhere.
In $VIMRUNTIME/colors/default.vim, `set bg&` sets bg=light, so let's
revert that as well.