feat(emacs): Evaluation settings for ORG blocks
Defines `rc/register-org-babel-languages` for specifying languages for
inline ORG code blocks. Takes 0 or more language specifiers; a
language specifier is a string containing the name of the
language (what goes in `#+BEGIN_SRC` declaration), preceded by '+' to
enable execution or '-' to disable execution. For example, to enable
`lisp` and disable `C` executions:

    (rc/register-org-babel-languages "+lisp" "-C")

If an entry is malformed, a warning is emitted. Valid entries match
the regular expression `/^[+-][a-zA-Z_-]+$/`. That is, '+' or '-'
followed by 1 or more letters, underscores, or dashes.